requirements.txt
kraken
pillow
streamlit
